Improved insertion time from 1h++ to 4min

+++ src/main/java/fr/diginamic/geoff/utils/Deserializer.java
@@ -44,11 +44,11 @@ public class Deserializer
      * @param part
      * @return
      */
-    public static double stringToDouble(String part)
+    public static Double stringToDouble(String part)
     {
         if (part == null || part.trim().isEmpty())
         {
-            return 0;
+            return (double) 0;
         }
         ;
         return Double.parseDouble(part);
